backend/skillmap-backend-starter/requirements.txt
fastapi
uvicorn
pandas
pdfplumber
python-docx
python-multipart
sqlalchemy
psycopg2
email-validator
passlib
python-jose[cryptography]
openai>=1.0.0
